* List_GEN.m ([TINDEX -setOffset:]): Cast count to int.
git-svn-id: svn://svn.savannah.nongnu.org/swarm/trunk@2562 84d386b0-d280-4341-8980-c4cb7e1eb7a5

+++ swarm/src/collections/List_GEN.m
@@ -578,7 +578,7 @@ PHASE(Using)
 
 - setOffset: (int)offset
 {
-  if ((offset < 0) || (offset >= collection->count))
+  if ((offset < 0) || (offset >= (int) collection->count))
     raiseEvent (OffsetOutOfRange, nil);
   link = (link_t) Start;
   position = 0;
